requirements.txt
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.2.0
click==8.1.7
dependency-injector==4.41.0
exceptiongroup==1.2.2
fastapi==0.112.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.7
jmespath==1.0.1
Mako==1.3.5
MarkupSafe==2.1.5
parso==0.8.3
passlib==1.7.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.37.2
typing_extensions==4.12.2
uvicorn==0.30.5
uvloop==0.19.0
watchfiles==0.23.0
websockets==12.0
